pyproject.toml
[tool.poetry]
name = "html-plot"
version = "0.1.0"
description = "Functions for creating HTML objects out of Matplotlib plots in Jupyter"
authors = ["kxmh42 <kxmh42@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "html_plot"}]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
python-box = "^7.0.1"
ipython = "^8.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
